Add Donkey Kong Arcade palette from Sho
In the previous update, I had forgotten to add this palette made by @ZReport for the Donkey Kong arcade port hack that has been making the rounds. Additing it to the root of the Miscellaneous folder as it also makes many early Nintendo releases have their arcade-like colors, so it's pretty valuable as an addition.